refactor(home): first-class demo/framework split of the home shell
The home tab was a 400-line demo dashboard hard-bound to a demo-only ViewModel
(4 demo-repo deps), so it could not be reduced to a compiling minimal shell by
in-place `// demo:begin/end` demarcation alone.

Split it cleanly along the demo/ convention:
  • feature/home/demo/HomeDashboard.kt   — the Money-Toolkit dashboard (moved from
    HomeScreen.kt): hero + loans carousel + bills/rates/exchange cards + tools grid.
    Removed its Scaffold/top-bar (that is the shell's job now).
  • feature/home/demo/ui/HomeViewModel.kt — the demo VM (moved), + its state models.
  • feature/home/HomeScreen.kt            — NEW framework shell: Scaffold + app-title
    top bar + settings action, always present. Renders HomeDashboard inside a
    `// demo:begin/end` block within a padded Box.

`customizer --clean` deletes feature/home/demo/ (via the **/demo/** glob) and strips
the demo blocks from the shell + HomeModule + the cmp-navigation callback chain
(homeGraph → authenticatedNavbarGraph → AuthenticatedNavbarNavigationScreen →
authenticatedGraph), leaving a home with a top bar + settings entry point and an
empty body for the fork to fill. No swap files, no hidden shells.

remove-demo.sh:
  • import-cleanup now also strips `kpt.feature.<any>.demo.*` (the shell's dashboard import)
  • dropped the obsolete step-7 shell-swap (home shell is first-class now)
