+++ feature/home/src/commonMain/kotlin/kpt/feature/home/demo/ui/HomeViewModel.kt
@@ -7,7 +7,7 @@
  *
  * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package kpt.feature.home.ui
+package kpt.feature.home.demo.ui
 
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.SharingStarted
